+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed addAmount to addAmounts
+- sumAmounts now expects an Amount[] parameter
+
 ## [1.0.0-alpha.2] - 2022-08-18
 
 ### Fixed